Break out pnet packets into their own crate
