+++ pnet_packet/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "pnet_packet"
+version = "0.19.0"
+authors = ["Robert Clipsham <robert@octarineparrot.com>"]
+license = "MIT/Apache-2.0"
+homepage = "https://github.com/libpnet/libpnet"
+repository = "https://github.com/libpnet/libpnet"
+description = "Cross-platform, binary packet parsing and manipulation"
+keywords = ["networking", "packet", "parsing", "raw"]
+categories = ["network-programming", "parser-implementations"]
+build = "build.rs"
+
+[features]
+default = ["with-syntex"]
+nightly = ["pnet_macros_plugin"]
+with-syntex = ["syntex", "pnet_macros/with-syntex"]
+
+[dependencies]
+pnet_base = { path = "../pnet_base" }
+pnet_macros = { path = "../pnet_macros", optional = true, version = ">=0.9" }
+pnet_macros_support = { path = "../pnet_macros_support", version = ">=0.1" }
+pnet_macros_plugin = { path = "../pnet_macros_plugin", optional = true, version = ">=0.1" }
+
+syntex = { version = "0.42", optional = true }
+
+[build-dependencies]
+glob = "0.2"
+syntex = { version = "0.42", optional = true }
+pnet_macros = { path = "../pnet_macros", version = ">=0.13" }
