+++ pnet_packet/src/udp.rs.in
@@ -6,8 +6,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use packet::Packet;
-use packet::ip::IpNextHeaderProtocols;
+use Packet;
+use ip::IpNextHeaderProtocols;
 
 use pnet_macros_support::types::*;
 
@@ -52,8 +52,8 @@ pub fn ipv4_checksum_adv(packet: &UdpPacket,
 
 #[test]
 fn udp_header_ipv4_test() {
-    use pnet::packet::ip::IpNextHeaderProtocols;
-    use pnet::packet::ipv4::MutableIpv4Packet;
+    use ip::IpNextHeaderProtocols;
+    use ipv4::MutableIpv4Packet;
 
     let mut packet = [0u8; 20 + 8 + 4];
     let ipv4_source = Ipv4Addr::new(192, 168, 0, 1);
@@ -122,8 +122,8 @@ pub fn ipv6_checksum_adv(packet: &UdpPacket,
 
 #[test]
 fn udp_header_ipv6_test() {
-    use packet::ip::IpNextHeaderProtocols;
-    use packet::ipv6::MutableIpv6Packet;
+    use ip::IpNextHeaderProtocols;
+    use ipv6::MutableIpv6Packet;
 
     let mut packet = [0u8; 40 + 8 + 4];
     let ipv6_source = Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1);
